Separate email into package
